requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
mlflow==2.11.3
python-multipart==0.0.9